Fix: change the basic structure of the project
