--- .prettierrc
@@ -1,8 +1,0 @@
-{
-  "trailingComma": "es5",
-  "tabWidth": 4,
-  "semi": true,
-  "singleQuote": true,
-  "printWidth": 120,
-  "endOfLine": "auto"
-}
