+++ client/.prettierrc
@@ -1,0 +1,8 @@
+{
+    "trailingComma": "es5",
+    "tabWidth": 2,
+    "semi": true,
+    "singleQuote": true,
+    "printWidth": 120,
+    "endOfLine": "auto"
+}
