+++ client/src/index.ts
@@ -1,0 +1,3 @@
+import './style.scss';
+
+fetch('http://localhost:3000/garage');
