+++ package.json
@@ -1,51 +1,35 @@
 {
-    "devDependencies": {
-        "@testing-library/dom": "^8.19.1",
-        "@testing-library/jest-dom": "^5.16.5",
-        "@types/jest": "^29.2.5",
-        "@types/node-sass": "^4.11.3",
-        "@typescript-eslint/eslint-plugin": "^5.46.1",
-        "@typescript-eslint/parser": "^5.46.1",
-        "@webpack-cli/generators": "^3.0.1",
-        "babel-register": "^6.26.0",
-        "css-loader": "^6.7.2",
-        "eslint": "^8.2.0",
-        "eslint-config-airbnb": "^19.0.4",
-        "eslint-plugin-import": "^2.25.3",
-        "eslint-plugin-jsx-a11y": "^6.5.1",
-        "eslint-plugin-prettier": "^4.2.1",
-        "eslint-plugin-react": "^7.28.0",
-        "eslint-plugin-react-hooks": "^4.3.0",
-        "eslint-webpack-plugin": "^3.2.0",
-        "html-loader": "^4.2.0",
-        "html-webpack-plugin": "^5.5.0",
-        "jest": "^29.3.1",
-        "jest-environment-jsdom": "^29.3.1",
-        "prettier": "2.8.3",
-        "sass": "^1.56.1",
-        "sass-loader": "^13.1.0",
-        "style-loader": "^3.3.1",
-        "ts-jest": "^29.0.3",
-        "ts-loader": "^9.4.2",
-        "typescript": "^4.9.4",
-        "webpack": "^5.75.0",
-        "webpack-cli": "^5.0.1",
-        "webpack-dev-server": "^4.11.1"
-    },
-    "version": "1.0.0",
-    "description": "My webpack project",
-    "name": "Clone kinopoisk",
-    "scripts": {
-        "build": "webpack --mode=production --node-env=production",
-        "build:dev": "webpack --mode=development --node-env=development",
-        "build:prod": "webpack --mode=production --node-env=production",
-        "watch": "webpack --watch --node-env=development",
-        "serve": "webpack serve --node-env=development",
-        "start": "webpack serve --open --mode=development --node-env=development",
-        "test": "jest"
-    },
-    "dependencies": {
-        "clean-webpack-plugin": "^4.0.0",
-        "ts-node": "^10.9.1"
-    }
+  "name": "rs-clone",
+  "version": "1.0.0",
+  "description": "",
+  "scripts": {
+    "server": "cross-env NODE_ENV=production ts-node app.ts",
+    "server:watch": "cross-env NODE_ENV=development nodemon app.ts",
+    "client:start": "npm run start --prefix client",
+    "client:install": "npm install --prefix client",
+    "client:build": "npm run build --prefix client",
+    "dev": "cross-env NODE_ENV=developmen concurrently \"npm run server:watch\" \"npm run client:start\""
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/Alesia-Abaeva/rs-clone.git"
+  },
+  "author": "",
+  "license": "ISC",
+  "bugs": {
+    "url": "https://github.com/Alesia-Abaeva/rs-clone/issues"
+  },
+  "homepage": "https://github.com/Alesia-Abaeva/rs-clone#readme",
+  "devDependencies": {
+    "@types/json-server": "^0.14.4",
+    "concurrently": "^7.6.0",
+    "cross-env": "^7.0.3",
+    "nodemon": "^2.0.20",
+    "ts-node": "^10.9.1",
+    "typescript": "^4.9.4"
+  },
+  "dependencies": {
+    "dotenv": "^16.0.3",
+    "json-server": "^0.17.1"
+  }
 }
