--- src/index.html
@@ -1,12 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-    <head>
-        <meta charset="UTF-8" />
-        <meta http-equiv="X-UA-Compatible" content="IE=edge" />
-        <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-        <title>RS Clone</title>
-    </head>
-    <body>
-        <div id="app"></div>
-    </body>
-</html>
