--- src/models/style-declaration.d.ts
@@ -1,4 +1,0 @@
-declare module '*.scss' {
-    const content: Record<string, string>;
-    export default content;
-}
