+++ tsconfig.json
@@ -1,14 +1,11 @@
 {
-    "exclude": ["node_modules"],
-    "include": ["src", "webpack.config.ts"],
-    "compilerOptions": {
-        "strict": true,
-        "noImplicitAny": true,
-        "noUnusedLocals": true,
-        "module": "commonjs",
-        "target": "es5",
-        "allowJs": true,
-        "moduleResolution": "node",
-        "esModuleInterop": true
-    }
+  "compilerOptions": {
+    "module": "commonjs",
+    "esModuleInterop": true,
+    "target": "es6",
+    "moduleResolution": "node",
+    "sourceMap": true,
+    "outDir": "dist"
+  },
+  "lib": ["es2015"]
 }
